Move sanitization tests into their own file.
Also adds an explicit test for [link](...) syntax to be sanitized.
